+++ packages/components/src/components/tabs/RouteTabs.vue
@@ -35,6 +35,7 @@ const itemContentClasses = computed<string>(() => tabsStyle.itemContent())
 <template>
   <TabsRoot
     :model-value="activeRouteName"
+    :style="props.styleConfig"
     :class="ThemeUtil.getClasses(themeProviderContext.theme.value, themeProviderContext.appearance.value)"
     class="tabs-default"
   >